Fix thresholds test by moving it to cmd/
Because the parsing of the threshold definitions now happens in cmd/, loadedTest.consolidateDeriveAndValidateConfig() specifically, this Engine test was completely invalid. You could add whatever values you wanted in its thresholds and it will have still passed...
