+++ core/engine_test.go
@@ -794,73 +794,6 @@ func TestRunTags(t *testing.T) {
 	}
 }
 
-func TestSetupTeardownThresholds(t *testing.T) {
-	t.Parallel()
-	tb := httpmultibin.NewHTTPMultiBin(t)
-
-	script := []byte(tb.Replacer.Replace(`
-		import http from "k6/http";
-		import { check } from "k6";
-		import { Counter } from "k6/metrics";
-
-		let statusCheck = { "status is 200": (r) => r.status === 200 }
-		let myCounter = new Counter("setup_teardown");
-
-		export let options = {
-			iterations: 5,
-			thresholds: {
-				"setup_teardown": ["count == 2"],
-				"iterations": ["count == 5"],
-				"http_reqs": ["count == 7"],
-			},
-		};
-
-		export function setup() {
-			check(http.get("HTTPBIN_IP_URL"), statusCheck) && myCounter.add(1);
-		};
-
-		export default function () {
-			check(http.get("HTTPBIN_IP_URL"), statusCheck);
-		};
-
-		export function teardown() {
-			check(http.get("HTTPBIN_IP_URL"), statusCheck) && myCounter.add(1);
-		};
-	`))
-
-	registry := metrics.NewRegistry()
-	builtinMetrics := metrics.RegisterBuiltinMetrics(registry)
-	runner, err := js.New(
-		&lib.RuntimeState{
-			Logger:         testutils.NewLogger(t),
-			BuiltinMetrics: builtinMetrics,
-			Registry:       registry,
-		},
-		&loader.SourceData{URL: &url.URL{Path: "/script.js"}, Data: script},
-		nil,
-	)
-	require.NoError(t, err)
-
-	test := newTestEngine(t, nil, runner, nil, lib.Options{
-		SystemTags:      &metrics.DefaultSystemTagSet,
-		SetupTimeout:    types.NullDurationFrom(3 * time.Second),
-		TeardownTimeout: types.NullDurationFrom(3 * time.Second),
-		VUs:             null.IntFrom(3),
-	})
-	defer test.wait()
-
-	errC := make(chan error)
-	go func() { errC <- test.run() }()
-
-	select {
-	case <-time.After(10 * time.Second):
-		t.Fatal("Test timed out")
-	case err := <-errC:
-		require.NoError(t, err)
-		require.False(t, test.engine.IsTainted())
-	}
-}
-
 func TestSetupException(t *testing.T) {
 	t.Parallel()
